New algorithm for 2D solid fractions computation (gerris--mainline--0.7--patch-11) gerris--mainline--0.7--patch-11 Keywords:
Does not use the gts_surface_inter etc... functions but a simple
computation of the intersections of the sides of the cell with the
surface, combined with a linear approximation of the piece of the
surface contained in the cell.

This is *much* faster, simpler and should be robust. It is also much
less picky about the degeneracies of the surfaces it can deal with.

It does not work yet for a varying level of refinement along the
surface.