Removed upstream self-intersection checks (gerris--mainline--0.7--patch-15) gerris--mainline--0.7--patch-15 Keywords:
This can cause the solid fractions algorithm to fail due to
consistency checks being violated by self-intersecting surfaces.

The messages will be a bit obscure. We'll fix that as we go.

The advantage is that in possibly a large number of cases with "minor"
self-intersections (e.g. nearly collocated vertices etc...) the
algorithm will proceed and should produce valid solid fractions.