GfsVariable are now events (gerris--mainline--0.7--patch-24) gerris--mainline--0.7--patch-24 Keywords:
These events are activated at the start of the timestep loop in run()
methods.

This means that the syntax for variables in parameter files has been
changed.

This involved a substantial cleanup and restructuring of the events
and their links with simulations.

The first application of this new mechanism is the new
GfsVariableResidual.