Changed computation of centered pressure gradients
This influences only the pressure gradient in mixed cells. It is now
computed as the surface-fraction-weighted averaged rather than just
the average. This is intuitively more consistent as this provides a
continuous interpolation when the face fractions vanish on one side of
the cell. More importantly this greatly stabilises the ocean model
near sharp headlands. It has a minimal influence on the convergence
rates of the test cases with boundaries (slightly improves the V
component and slightly degrades the U component).