Bug fix for GfsEventStop
EventStop was using a temporary variable to store the previous timestep values.
This variable was not initialised when adapting the mesh (temporary variables
are not initialised by default).

This has been fixed by using a "real" variable instead but with a NULL name.

This meant allowing variables with a NULL name i.e. hidden permanent variables.