"Non-hydrostatic pressure" implementation
This is useful for two-phase flows with a (flattish) interface, variable
density and gravity. The hydrostatic pressure is substracted off and the
remaining pressure jump across the interface due to gravity is discretised
using the same technique as for surface tension. This guarantees an accurate
solution free of spurious (gravity) currents.