Pressure-correction in the Crank-Nicholson scheme is turned off by default
This was degrading the results for a number of test cases (capillary waves in
particular). The reasons why are still unclear.

Note that this does not mean that this version of the code is the same as the
version prior to the "Pressure term is included in RHS of viscosity solve"
patch, as this version also includes an important bug fix for the advection
terms when surface tension is present.

The pressure-correction Crank-Nicholson scheme can be turned off explicitly
when required (e.g. "hydrostatic" test case).