gfs_get_poisson_problem() returns a linear system with a unique solution Ignore-this: 27031b757e1259e01ca324a9f7359ccf
For Neumann bcs everywhere, the solution of the Poisson problem is
defined to within a constant. This causes convergence problems for
external linear solvers such as hypre. The solution is to remove the
extra degree of freedom by setting the value of the solution in a
single row. Note that the check for Neumann bcs assumes a simply
connected domain topology.