gfs_interpolate() uses 5 points interpolation. Ignore-this: 7aeb604d4bf5815de57308354a8900a5
This doesn't smooth as much as 4 points interpolation. Note that the
size of the array filled by gfs_cell_corner_values() as grown by one
element as a result of this change. It is imperative to adjust the
size of the array for all calls made to gfs_cell_corner_values().